created & ran migration for messages, notifications and added new cols to users table, return property owner details to api

+++ api.solidagents.com/database/migrations/2021_12_25_132727_create_likes_table.php
@@ -13,13 +13,15 @@ class CreateLikesTable extends Migration
      */
     public function up()
     {
-        Schema::create('likes', function (Blueprint $table) {
-            $table->id();
-            $table->bigInteger('user_id');
-            $table->bigInteger('property_id');
-            $table->foreign('property_id')->references('id')->on('properties')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('likes')) { //this was the fix for an error when I was told that the table likes already exist during migration
+            Schema::create('likes', function (Blueprint $table) {
+                $table->id();
+                $table->bigInteger('user_id');
+                $table->bigInteger('property_id');
+                $table->foreign('property_id')->references('id')->on('properties')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
